Rework devstack-jobs templating
Currently devstack-jobs template creates jobs
gate-X
periodic-X-master
periodic-X-liberty
periodic-X-kilo
gate-X-liberty-nv
gate-X-kilo-nv
gate-X-kilo
gate-X-liberty

But only a few of the jobs are used in periodic pipelines or in stable
branches for branchless tempest.

Create new job groups devstack-periodic-master-jobs and
devstack-stable-jobs that can be used for the periodic and stable
branches.

This reduces the number of generated jobs from 8744 to 8382.

Change-Id: I4782f0d36965a0edf1733021be100f0a79ab5534