Templatize experimental bindep jobs
The experimental bindep-using nova Python 2.7 unit test job is built
from a job-template now, updated for parity with the current state
of its traditional non-bindep equivalent, instantiated for nova via
a job-group, and added in the zuul layout with a project-template.
This is a preparatory step toward adding more bindep-based job types
and on a broader number of projects before we plan to cut over
official jobs to new non-bare/minimal DIB-based nodes.

Change-Id: I4d5c96d02c3d06c71f32cb3bea8ae3dde699a86b
Signed-off-by: Jeremy Stanley <fungi@yuggoth.org>