Add distil jobs back
Distil project has been evolved internally in our company, we use it in
production, and we want to keep it upstream first, following the
openstack way. So, adding the gate jobs back is just the first step.

The py27 jobs result: http://dpaste.com/336AMP4(will expire in a month)

Change-Id: Ib475cdec4246920b279c40de31e68960b990b4b6