Add javascript jobs to stackviz
Stackviz is a tempest runtime visualization tool that serves up a
static website with test results. While most of the data processing is
done via a python module, a large amount of the codebase is written in
AngularJS. Javascript jobs will be necessary to ensure the codebase is
clean and maintainable as work progresses and features are added.

Change-Id: Ie14fae7fbf70c37ce0b53fe54ba54884b1af16da