Make postgres job non-voting on Ironic and switch it to IPA
As part of our gate cleanup we agreed to reduce number of jobs voting
on ironic currently. Postgres job is nice to have, but it adds to random
failures we experience. So we make it non-voting.

We're also removing support for the old ramdisk, so switch this job to IPA.

Change-Id: I40cbc6ef3512f30a14e87bb602ea4ea2bcaf61ae