Switch most of ironic jobs to iPXE and remove a separate iPXE job
On the ironic midcycle we agreed to reduce number of jobs voting on ironic.
We also agreed to switch to using iPXE in more jobs, as it's expected to be
used in production more often than plain PXE.

This patch makes all pxe_ipa jobs use iPXE and all agent_ssh jobs use PXE.
Thus a separate pxe_ipa-ipxe job is removed from the gate.

Gate jobs on ironicclient and ironic-lib are switched to iPXE as well.
The pxe_ssh job testing the old ramdisk is left as it is while we're still
deciding its fate.

TODO:
* Switch ironic-inspector jobs to iPXE as well as soon as its devstack plugins
  supports iPXE.
* Rename ironic jobs to better reflect what they do.

Change-Id: I9d94bfcbe3bb46b81734414e3f67ba4dde61725a