Make Ironic VM memory and count configurable
TinyIPA shouldn't require as much memory to run as the coreos ramdisk,
this patch makes the Ironic fake baremetal VMs memory amount
configurable per job so that the tinyipa job can be configured with less
memory. This patch also makes the number of Ironic VMs configurable,
laying the foundation for enabling multiple in the future.

Change-Id: I9a96f6c5f48de2007f2d94607007ae49527bf87d