Enable coverage non-voting for networking-cisco
Goal is to provide a tool for reviewers to better tell if unit tests are
exercising the code being added/changed, in a effort to (indirectly)
improve quality. Prior to this commit coverage was 61%.

If any projects are also doing automated comparisons of coverage results
with a baseline, please let me know, as we would be interested in
exploring taking advantage of that capability too.

Change-Id: Ia2b32b25b27e4a8b4efa60439dc69e80cf49f548