Recreate workspace in zuul-cloner cleanup
In cases where zuul-cloner is aborted during a git clone operation,
git will remove the git work tree in its cleanup.  The work tree
in these jobs is the workspace directory, which means that subsequent
jenkins post-build actions can not run because the workspace has
been removed.

To reduce the likelyhood of this having an impact, recreate the
workspace directory if needed during the cleanup trap.

This is still slightly racy (there may be a brief window where
there is no workspace), but nonetheless, this should work in most
circumstances.

Change-Id: Iab2257e622591f936bf3d0619c29fa4c7b569bd6