Add experimental-bindep-jobs to other-requirements users
Add experimental-bindep-jobs to all repositories that have
an other-requirements.txt file in their projects. This allows to test
that they continue to work if we switch to bindep based jobs.

Note that experimental-bindep-jobs includes more jobs than setup for
some of these repositories - but since those are experimental, this
should be fine.

Repositories that have other-requirements.txt are: requirements,
release-tools, tripleo-incubator, and bindep.

Change-Id: Iabb59e698d0d4e5912885b10114233b2b9ccffcd