Add broadview-ui project
broadview-ui is a Horizon-based panel that can be used to configure
openstack/broadview-lib and openstack/broadview-collector.

The UI allows an administrator to select which metrics supported by
broadview-lib are to be collected, and how that collection is to be
performed.

This project is the third of an overall contribution that will supply
BroadView metrics data gathered by broadview-lib to monitoring
projects like OpenStack Monasca.

Change-Id: Ia97960ab02647daa505a19e5d49a777a3f3ba83f