Add ipmitool non-voting test to Ironic
The commit 7ffa1d58ce9fd9e32cdab142a1f12feae38c3097 added support to
configure nodes with ipmitool drivers in DevStack ironic to run with
virtual machines. The main goal of that patch was to allow us to test
the production ipmitool drivers in the gate replacing the *_ssh testing
drivers in the future.

This patch is adding two new non-voting jobs to Ironic: pxe_ipmitool and
agent_ipmitool.

Related-Bug: #1544642
Change-Id: I9af881a06ceb2d9a2133db4cf2dc3ffceb0813ff