Add scenario003 in Puppet OpenStack CI
Puppet OpenStack CI jobs are currently deploying 2 scenarios [1].
We start to have serious performances issues, because of the number of
services that we run on each scenario.

Having a third scenario would help us to balance the services installed
on each scenario and reduce CI failures, mainly due to job timeouts.

For now, scenario003 would be non-voting since we need to write a
manifest before and make sure it's more stable.

[1] https://github.com/openstack/puppet-openstack-integration#description

Change-Id: I57dabc25aed4faec5ebfc22f23f7df004b64e9a7