Add diskimage-builder-release group
We have moved to a situation where diskimage-builder is being used
widely outside of a strict tripleo setup, so add a new release group
so people can have permissions just to release dib.  We will add
tripleo-release to this group to maintain status quo.

While we are here, bring dib-utils (diskimage-builder's util library)
into line with dib by using the same acl config.  tripleo-core is a
member of this so no permissions are removed.

Change-Id: I7287157b7dec680d7d5c44a075ef092ee0f3a6d5