Add openstack client command overlap checking
With the growing number of openstack client plugins, a large list
of common objects, and the fact that any varity of plugins might be
loaded, it would be hard to stop a command overlap issue before it
get's merged. So, I took a stab at adding this to see if others
think it should be part of the command checking gate job.

Change-Id: Ifb36cb1211283987959e4cb488c5f62ffeb8ff06