Make tweaks to neutron-plus job
This change moves the job definition closer to the other
job definitions available in the neutron.yaml file.

It also replaces the suffix with 'scenarios', which is more
intuitive than 'plus', and better reflects the nature of testing
that this job is going to carry out.

Finally, it adds a post hook, and tune some configuration knobs
to allow for finer control of its execution.

Change-Id: I24be6ddc14a40950bb9ffc2c91e6150a6a74d266