Remove gate-openstack-ansible-commit-nv job
The gate-openstack-ansible-commit-nv used the ubuntu-trusty image and was
implemented as a tester based on the advice that all current dsvm jobs
would be moved to this image.

As this job fails most often and there clearly is no major move happening
at this stage, we'd rather remove this job for now to reduce nodepool usage
and reduce delays caused by waiting for nodes with this image to be available.

Change-Id: I425c8ee6a214d7fef3419b96a94b0f55b24f8aef