Change kolla gate timeout to 90 minutes
Our centos gates are often timing out because we do not
have a local mirror of centos in the infrastructure.  While
I'd like to fix that problem (to match the Ubuntu mirrors we
do have) our gating fails at about 90% complete most of the
time on centos deployment.  The problem isn't memory related
but related to the booting of the 1 VM we do (virt on virt).

if 90 minutes isn't acceptable, we could try 75 minutes.  We
are pretty close to having a full centos deploy in the gate
complete.  See:
http://logs.openstack.org/71/286771/1/check/gate-kolla-dsvm-deploy-centos-source/f552bf8/console.html

Change-Id: Icc416f08d6b27cda58a44dd5b9d93a3d6262a00a