Add experimental job to test xstatic packaging
We need a way to test if a particular change
(I14e0b5986d3433afd796ea44e66dfafd5003e732) will allow wheels to be
built with setuptools_scm.

This job is a temporary clone of {name}-tarball to put the build
artifacts on the log server so we can check if it works rather
than trigging a real build (and hence upload to pypi).

This should be removed after builds with the above change are working.

Change-Id: Idb36459bcdfd85d7765e849dfda105437bb8bd0c