Run full tempest on CentOS/Fedora
In I3a58144426ff4e904c316210b647379b1ba7305d I managed to set
DEVSTACK_GATE_TEMPEST_FULL=1 *before* the smoketest definition, and
thanks to the if-gate in devstack-gate this means I basically switched
the job to running full tempest tests.

This was on my todo list to implement in a more controlled fashion
(last time I tried it failed), but luckily it all works.  So since
we're running full tests anyway, let's clean this up.

Change-Id: I827b66e3070931fadfc6d2ab7437cee8e0300159