Remove gate-ironic-dsvm-functional from experimental queue for devstack
On master (mitaka), Ironic has a devstack plugin in-tree, so the
job fails on master devstack changes.

So the only point of this job for devstack would be to test devstack
changes in lib/ironic on stable branches. We can't filter the job
to only run on stable branches though, because it's also in Ironic's
experimental queue (which should work on master (mitaka)).

This just removes the job from devstack's experimental queue since
it doesn't work for master and any stable branch changes that care
will be tested by gate-tempest-dsvm-ironic jobs on the Ironic repo
itself.

Change-Id: I8ac54c682e7bf1aada9e7c4c1c0278fdd3378dd9