Remove pypy's gate for zaqarclient
This gate has been non-voting for quite a bit and I'm not aware of
anyone that's working on fixing it. The Zaqar team can't dedicate
resources to fixing this gate and removing it would help saving some
cycles in our gate. Besides, zaqar client is depending on keystone
client and openstack client, both of them don't support pypy.

Change-Id: If454d14665b1ba1aca61112c7f25e679967473eb