Remove gate-tempest-dsvm-full-ceph job
Due to bug 1551305 the gate-tempest-dsvm-full-ceph job
is non-voting.

The long-term goal is to switch over to the
gate-tempest-dsvm-full-devstack-plugin-ceph-nv job and
make it voting since it uses upstream packages rather
than what's in the distro (which is our problem with
ubuntu trusty and bug 1551305).

Also, since both jobs are non-voting, we're wasting
nodes in cinder/glance/nova running both, we need to
just drop the gate-tempest-dsvm-full-ceph job since
we have the same coverage with the plugin job.

Note that os-brick and glance-store are using src
job variants of gate-tempest-dsvm-full-ceph which
will need to be replaced with src job variants of
the ceph plugin job, which will come in a later
patch.

Change-Id: I44c3a0c35e8dbcc49c0ac1800e9075cc3a169fdd