Add non-voting diskimage-builder jobs
This commit allows the definition of non-voting gate jobs by defining
a new job-suffix. 'extras' was added as a new job pre-suffix because
just 'nv' implies we would be testing the same things as the
non-postfixed variant.  This was done to facilitate adding a Gentoo
non-voting gate job, also in this commit.

Change-Id: I08f35397a4cce05b90d88077922dc38ece5c2b55
Depends-On: I0d2768a912b350e1b2cf40d2fd227fdc767b1bcb
Signed-off-by: Matthew Thode <mthode@mthode.org>