Move puppet jobs to ubuntu-trusty
This requires quite a bit of refactoring due to the way the puppet jobs
were organized previously. Now instead of assuming we are using subtypes
of a devstack node we assume nothing about the os-type and are very
explicit. This allows us to continue to use devstack-centos7 while
replacing devstack-trusty with ubuntu-trusty.

Change-Id: I1e8bb4c8417813caaad83a26523dca4107726d92