Add devstack-plugin-additional-pkg-repos
This devstack plugin will allow for testing a variety of system-level
tools within a devstack job.

The original idea comes from needing to test libvirt+qemu in nova but
this plugin can trivially be extended to allow neutron to test newer
version of ovs or dpdk.

Currently it is not attached to any Big Tent project but once it's
utility is proven I imagine it will fall under either Nova or QA.

Change-Id: I5a3f3ca671b3cf25d71783dbb197e4449cac3e34