Adding searchlight-ui repo and jobs
The Searchlight content from Horizon is moving from a patch
on master to be developed as a plugin to Horizon. This change
sets up the repository and the gate jobs for the new repository
so the code can be moved.

At the horizon mitaka mid-cycle, we decided to move this to a plugin
for the Mitaka release. It was not possible to do so until
several patch dependencies merged on horizon master. Those
have now merged, so we are able to put this into its own repo.

Created temporary repository from the review on master.

On top of that, considerable moving of files and additional
pieces were added to complete the plugin.

https://pypi.python.org/pypi/searchlight-ui has been created with
the appropriate permissions for openstackci.

Change-Id: I528537a525a54db3887688ab49b33983403dd2ca
Needed-by: I035c479102200cce38511482f433b030b066584d