Add element to handle generic package installs
We end up needing to install a fair bit of stuff on top of the minimal
images in order to make them useable. Do this in a generic infra package
needs element that can grow to accomodate our needs.

For the future we can hopefully get some of this into dib itself so that
others can easily build instances they can ssh into.

Change-Id: I5826df829971e182cd0713b29df1cf70c119a0ab