Add mitaka stable branch jobs
This covers the periodic bitrot and tempest jobs for mitaka as well as
the mitaka specific jobs for devstack-gate and tempest to test against
their master branches as they do not have stable branches.

This should be safe to merge before every project has a mitaka branch
because cloning should fall back to master if a project does not have a
mitaka branch yet.

Change-Id: I7bb0f9aa53fb7fee07d70f32b92f4fd6f4eabd46