Add a microversion-parse project
Modern microversion headers can be tricky to parse so a group of
concerned citizens decided we need a small library to parse them
in a consistent fashion. This adds a project, microversion-parse,
that contains a python library to do so. It's upstream is at
https://github.com/cdent/microversion-parse

For this project, no distinction is made in gerrit between core,
release, or stable maintainers. Core can do any of these.

Change-Id: If617d1ccf55e469a7ccd5647ab7358436971fd00