Exclude gate-tempest-dsvm-ironic-pxe_ipa from stable/kilo
gate-tempest-dsvm-ironic-pxe_ipa does not run reliably in
stable/kilo. It's meant as a replacement for the
gate-tempest-dsvm-ironic-pxe_ssh job starting in liberty,
so let's just exclude the pxe_ipa job from kilo since it's
blocking our ability to land ironic changes in kilo. We
should have adequate coverage with the ssh job.

Change-Id: I1453fdf3bfbb978a6a484f389827642c895f18e6