Give ironic tempest-full job more "bare metal"
This switches the ramdisk to tinyipa, which requires less RAM to run. As
such, it reduces the RAM allocated to the VMs to 384 (the minimum
observed amount of RAM for tinyipa to run, and spins up 7 of them
instead of one (based on a quick analysis of dstat on some gate runs).

NB this job doesn't pass right now :)

Change-Id: Ib9df614f2c70a1ffee4a5aadc1c9684f9a11451d
Partial-Bug: #1561760