Rename openstack-ansible-ironic to openstack-ansible-os_ironic
In order to ensure that all OpenStack Service roles are named with
the same convention (using the 'os_' prefix) and therefore allow
the deployment tooling to react to the roles in the same way for
production wheel repo building, we ask for the rename of the
OpenStack-Ansible Ironic repository.

As this deliverable is important to have as part of our Mitaka
release, we'll be tagging prior to the Newton summit. As such we'll
be using the GitHub URL for this repo (it'll handle the rename
redirect) in all code that refers to it to ensure that the tags can
build before and after the rename.

That said, it would be preferred to handle the rename prior to the
first tag so if it's possible to get it done prior to the summit
then we can consider holding off doing the tag until the rename is
complete.

Needed-By: I2ab80e4fdb413d469c7aa65e5f018b66cc203b7c
Change-Id: Ieeff38d029c592ced7a1ab4bc1af42f3ff1e9fcd