Watcher multinode - disable plugin in subnode
The Watcher multinode experimental job is currently not working
due to the fact that both the Watcher and Ceilometer plugins are
enabled onto the subnodes. Hence, this changeset introduces the new
DEVSTACK_SUBNODE_CONFIG variable to avoid the enabling of plugins
that only are meant to be activated on the primary node.

Change-Id: I6aea168a5d2ad61ed4ad36b25de8faeee044c543