upload both sdist and wheel on pre-release
We used to only upload wheels of pre-release packages to protect folks
who were using old versions of pip from having those things installed
"accidentally". It has been long enough, and we expect everyone to have
upgraded their pip by now, so update the job definitions so that when we
tag a pre-release we upload both the sdist and the wheel.

Change-Id: I7c1563b130ba5640f8d47c8795be64238532f56c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>