Request for new repo - python-k8sclient
Split out Kubernetes client Magnum into a separate project. First
consumer is likely to be Magnum. We will be adding Magnum core
to the ACL so they can maintain it if they choose to. Please
add "dims-v" as the sole core member of the group. The idea is
to encourage other folks to contribute to this library as well
as allow Magnum cores a full say in the repo.

NOTE: this is not part of any official project at the moment,
so there is no governance review for this package.

Change-Id: I1472e631ec7f06e74e73c27edd68a1e2dc7b23e1