Add openstack/monasca-transform project
The monasca-transform project is a new component in Monasca that
will aggregate and transform metrics. For example, it will consume
metrics for Swift from the Monasca metrics topic and count the total
number of times a container or object is accessed and group them by
project. Then it will publish the aggregated metric to the Kafka
metrics topic where the rest of the existing Monasca components
will be used to store, alarm and notify.

Change-Id: I7826768e174a229bc184ea38d75d212ac957e26c
Needed-By: I9227a6afb468e05aaeb939930b872d1c64f79551