Execute sbuild after relogin
Currently, pkgos-infra-install-sbuild is launch directly after sbuild is
ran. This leads to this error:
"User jenkins is not currently a member of group sbuild, but is in the
system group database You need to log in again to gain sbuild group
privileges"

So this patch splits the single shell script into 2, so that it does
the "log in again" part described in the above error message. This
should fix the problem.

Change-Id: Idf35d7c143e2fb67ccf6cdcefce74c3d45e9a49c