Call npm prune before npm shrinkwrap
There is a bug(?) in npm's shrinkwrap feature, which will throw
exceptions if any included dependency contains an npm-shrinkwrap
file that includes dev dependencies. This currently impacts
tripleo-ui, throwing the following errors:

http://paste.openstack.org/show/493316/

THe issue is known, and unresolved. It's available here:

https://github.com/npm/npm/issues/6298

This fix ensures that `npm prune` is invoked first, which
explicitly removes any dev dependencies from the
dependency tree. Since we need these for any previous
commands, I've also moved it to the end of the job execution.

Note that the functionality of shrinkwrap varies by version. We
will have to revisit this when testing the next Node LTS.

Change-Id: I3c9b17e35a4fe64ede8248c267cf3a7ae927bc54