Update Gerrit ACLs for vendor specific networking-* projects
As agreed in commit 846a52873, vendor specific drivers are beyond
what the Neutron team can oversee. As result, release management
and stable backports are delegated to the respective release and
core teams, who are in the best position to assess when and how
to tag/branch, and whether defect fixes are not a risk to the
stability of their solutions (if they have stable branches).

Change-Id: I98fa3e5d88a638610227c044d20eca5f6c58de64
Depends-on: Icb3d33af0e32919bb13c52b3662f481c2c99011c