Translations: Temporarily stop master sync for python projects
To allow translation of release notes, we need to change the srcdir
argument for zanata. The locale files stay the same, just the way they
are referenced is different now.

First, we disable syncing of master for projects with one module
(most projects, exceptions include horizon and keystone) temporarily.

Then, a manual update of resources is done.

And finally syncing is enabled again.

This change disables syncing and lays the ground work for how the setup
will be later:
* stable/mitaka will use old setup.
* master for *all* projects will use same settings, so this unifies
  current django and python projects, independent of number of modules.

Change-Id: I9719e256bbd213d9b9265faaf4e4e32632c5d8d7