Enable magnum plugin on shade and use our atomic generated images
This patch had been proposed before:
https://review.openstack.org/#/c/295969

But had to be reverted because we were consuming external images
from fedorapeople, that caused our tests to timeout. We are now on
the situation where we have our own custom atomic images, to start
adding that on shade as a previous step for adding magnum integration
to shade.

Change-Id: Ifae8e54e78f6f36cf025ac92ba704391b2052769