Make the ceph devstack-plugin-based job gating on Nova
This gets us back to gating on working ceph bits. The job has been
stable for quite some time now (over a month) and the only ceph-specific
failures have been actual regressions we have let into the tree because
this is not gating. We know this because we have a team of nova and
cinder people now monitoring status on a daily basis.

Change-Id: If97adae3fdbb823f0c7f6c01da96355afede4605