puppet: use gerrit-git-prep for all jobs
Stop using zuul-clone macro because the other-requirements.txt was not
found by install-distro-packages.sh script, since it looks at the root
of the workspace.
By using gerrit-git-prep, we have the repo directly at the root of the
workspace, other-requirements.txt will be found and we'll unblock our
CI, that is currently broken because some packages are installed by
default bindep packages.

Change-Id: I5be52b5790655d4361817890f2bf4af382924e2c