run-wheel.sh: only run wheel for Python projects
tarball JJB macro run wheel, but we have to verify if the project is
whether or not a Python project.

This patch will make possible for Puppet modules to run the tarball
macro without complexity.
We check if it's a Python project if metadata.json file IS NOT present.
Indeed, this file is really Puppet specific.

Change-Id: Id2e307ae53b8ef1bb2be71c8ba3541cadb5ca12b