Stop uploading to job logs to swift
We have some headroom on the logs filesystem and this allows us to go
back to more reliable methods of grabbing logs. Specifically we are not
required to have http access back to the jenkins master from every job
that runs into order to get the console logs then upload them to swift.

Change-Id: Ie8479c224b2fd3c3efca4e1bf4157656540eccff