Implement higher job timeout for Ansible role jobs
Some of the OpenStack-Ansible (OSA) roles construct enough of
an OpenStack environment and execute enough functional testing
for the job to easily reach the 30 min default time limit.

This patch implements a higher limit to cater for these jobs
to ensure that OSA can properly execute functional testing in
the roles without having to compromise o quality to accommodate
a short job timeout.

Change-Id: I780b735de54c23cfea14ddafa8ad39aff16e92eb