Make two Trove scenario tests part of the check
This change proposes to move two of the Trove scenario tests that were
being run in experimental to the check. These tests pass, but not
reliably. The failures are now largely related to timeouts and these
are being addressed in[1], other false-negatives which were seen
earlier have been now addressed.

Since these are now in the check (and no longer experimental) the
branches where they will run has been restricted to master, and moving
forward, newton, ocata, etc., The stability of the tests in mitaka is
not sufficient to make it worthwhile running there, and the tests are
not complete in earlier releases.

[1] https://bugs.launchpad.net/trove/+bug/1571092

Change-Id: I9135544e5c97ad9f129945006cc5b16a70d617e4