Enable g-api and n-api for glance_store functional
The check to test if glance_store should be installed from git
is under install_glance in the lib/glance. This is only called if
"is_service_enabled g-api n-api" is true, so these need to be in the
ENABLED_SERVICES.

This should solve the "The following LIBS_FROM_GIT were not installed
correct:  glance_store" error.

Change-Id: I33e3507c818d3f65242e11276f951b64a7da0cb0