Add job to build generic debian packages
Now that openstack-pkg-tools is built and publish, this job uses it
to build other packages.

To actually test this job, this patch also adds the deb-spice-html5
package, which is a very trivial package with no dependency, in order
to test building with this patch. I'm not sure it really is needed
to maintain the spice-html5 package on Gerrit, so this may be just
temporary. Other Git repositories for all of the other packages will
be added once this package builds fine and is published.

Also, as the Jessie image doesn't contain the debian-openstack repo
yet, we add it to the sources.list. This will be able to go away
next week, according to pabelanger.

Change-Id: I15660191360ed68fb1fb5a3a4170ad2353336ebf
Needed-By: I1188b90b3d2039f712fb1efda69e2a6157621079