Add KMIP Experiment Check to Barbican
Barbican currently supports a plugin architecture for secret stores.
One such secret store implementation is the KMIP secret store. There
are no checks in place at the moment to test the functionality of
Barbican with this plugin.

This patch creates a new Jenkins job to test Barbican with the KMIP
secret store plugin. The end effect will be running all of Barbican's
functional tests using a KMIP secret store plugin. This change enables
the PyKMIP server to run as part of the devstack setup.

The tests can only be run on the latest code. The older stable
branches for kilo, liberty, and mitaka are not supported because the
necessary devstack plugin code is not backported to those branches.
The KMIP secret store job is experimental for now until tests verify
the correct operation.

Change-Id: I3fc6abe6f34bcb31c9ea01bfd7c7ccfd22b28bcc
Depends-On: Ib804fa97545f14ed866bfd73bb251e85923a2e4e