Use setfiles to restore unbound selinux policy
We finalise the unbound install at run parts level 99, but the dib built
in selinux restore runs at level 11 so we need to handle this directly
for unbound.

We switch from restorecon to setfiles becaues restorecon does not do the
right thing in a chroot (it is chroot aware and essentially noops for us
because we are in a chroot).

Change-Id: I7af6a4925e191a7d1bc78fb2d66d9ce74b944a39