Fix comment on requirements job
Currently failures of this job leave a broken comment that doesn't
match the test-result regexes, meaning that you can't see the result
in the callout box.  See [1].

I863d88e9a7ed2fd41924b8fc4a12dbea3ee2b205 fixes the root cause, but
this message is still a bit wrong because it doesn't have the
[PASSED|FAILURE] bit.

I don't really see why we need this special result at all.  If it
fails, you have to read the log anyway to figure out what is wrong, so
move the message pointing to the docs to where the failure actually
happens in the logs.

[1] https://review.openstack.org/#/c/292193/

Change-Id: Ie5e7d2318a510a5811c9131d56594e856da9dcc5