Add ci-cd-pipeline-app-murano project to openstack
During summit session with Murano team there was decision about developing
group of applications separately and renaming old repository murano-apps to
example-murano-apps. Current repository will be used for developing
group of Murano applications, which allow to create easily reproducable
ci-cd pipeline. It will be some separate applications (Jenkins, Gerrit)
and also all in one applications (Jenkins+Gerrit).

The purpose of current project is to develop and maintain applications
mentioned above. So this project is not a Murano development, it's about
development specific applications, which can be used by Murano.

Change-Id: Iabf23f13f2faed3cd70d95c945c55d9abccba430